[Android] SetUnhandledSignals added to CrashpadClient
Chromium change:
https://chromium.googlesource.com/chromium/src/+/1c92945c0e3457c11f50ec4b6bc620163c030fe4

crash: don't handle browser SIGQUIT on Android.

On Android, SIGQUIT is masked by default and handled specially by ART to
generate stack traces on demand (usually during an ANR or bugreport) in
a way that does not intrinsically terminate the process (though the
process may be killed afterward in some cases).

Apps that use WebView sometimes unmask SIGQUIT in the browser process
due to having their own crash handler or similar, and if that happens,
we don't want to treat this as a browser process crash and generate a
minidump, as there's no real way to know why the signal was sent.

Just disable handling SIGQUIT on Android in the browser process. This
leaves child processes alone; they only run our code so can't end up
with SIGQUIT accidentally unmasked.

Fixed: 1077890