[Android] Changes in FeedSurfaceCoordinator
Chromium change:
https://chromium.googlesource.com/chromium/src/+/e274ab4bfbb68201b4c2cf36e8b7c0cf29981c17

Add enable_feed_v1 compile flag

And restructure feed code so that we can use it.

This flag pulls Feed v1 out of the build.
It's a bit different that enable_feed_v2. enable_feed_v2
retains most of the v2 code in chromium -- only the internal
code is actually removed when enable_feed_v2 is off.
If both flags are ON (the default), we use the InterestFeedV2
feature to select which version to use. To facilitate this,
only feed_feature_list.cc and FeedFeatures.java should inspect
the state of this feature directly.

If both build flags are OFF, we use the V2 UI classes as
essentially a pass-through to a recycler view, but will
not request any Feed content.

The compile flag enable_feed_in_chrome was removed.

Other changes:
* I've moved some files around to make it more clear what code
  is for v1 only.
* I've added FeedV1 and FeedV2 java classes to wrap access to
  version specific things, so that FeedV1 can be removed using
  the compile flag.

Bug: 1129187